Split intersection into its own module
